Amélioration des datatables étudiants avec boutons d'action visibles et édition en ligne
Modifications principales:
- Ajout du partial student_action_buttons.blade.php pour les boutons d'action visibles
- Boutons Voir (bleu), Modifier (orange), Supprimer (rouge) directement visibles
- Menu déroulant 'Plus' pour les options supplémentaires
- Mise à jour de inline_editing.js avec support responsive et visibilité colonnes
- Ajout du CSS datatable_responsive.css pour le style moderne des datatables
- Modification de list.blade.php et list_all.blade.php pour utiliser le nouveau partial

Fonctionnalités:
- Edition en ligne des cellules (clic pour modifier)
- Boutons d'action toujours visibles
- Option de visibilité des colonnes
- Export Excel/PDF des colonnes visibles uniquement
- Support responsive pour mobile

+++ database/migrations/2025_08_25_100000_create_encaissements_table.php
@@ -17,17 +17,17 @@ class CreateEncaissementsTable extends Migration
         if (!Schema::hasTable('encaissements')) {
             Schema::create('encaissements', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedBigInteger('student_id');
-                $table->unsignedBigInteger('payment_id');
-                $table->unsignedBigInteger('payment_record_id');
-                $table->unsignedBigInteger('class_id');
+                $table->unsignedInteger('student_id');
+                $table->unsignedInteger('payment_id');
+                $table->unsignedInteger('payment_record_id');
+                $table->unsignedInteger('class_id');
                 $table->enum('type_encaissement', ['ADRA', 'TEAM3']);
                 $table->decimal('montant_original', 10, 2);
                 $table->decimal('pourcentage_pris_en_charge', 5, 2);
                 $table->decimal('montant_encaisse', 10, 2);
                 $table->date('date_encaissement');
                 $table->string('reference_encaissement')->unique();
-                $table->unsignedBigInteger('created_by');
+                $table->unsignedInteger('created_by');
                 $table->string('year');
                 $table->text('observations')->nullable();
                 $table->timestamps();
@@ -48,7 +48,6 @@ class CreateEncaissementsTable extends Migration
                 $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
             });
         }
-        // If table exists, skip creation - indexes will be handled by Laravel automatically
     }
 
     /**
