+++ database/migrations/2025_08_25_100001_create_recettes_table.php
@@ -18,24 +18,24 @@ class CreateRecettesTable extends Migration
             Schema::create('recettes', function (Blueprint $table) {
                 $table->id();
                 $table->date('date_recette');
-                $table->unsignedBigInteger('student_id')->nullable();
-                $table->string('beneficiaire_nom')->nullable(); // Pour les recettes non liées à un étudiant
-                $table->unsignedBigInteger('class_id')->nullable();
-                $table->unsignedBigInteger('payment_id')->nullable();
-                $table->unsignedBigInteger('payment_record_id')->nullable();
-                $table->unsignedBigInteger('encaissement_id')->nullable(); // Référence vers encaissement si applicable
+                $table->unsignedInteger('student_id')->nullable();
+                $table->string('beneficiaire_nom')->nullable();
+                $table->unsignedInteger('class_id')->nullable();
+                $table->unsignedInteger('payment_id')->nullable();
+                $table->unsignedInteger('payment_record_id')->nullable();
+                $table->unsignedBigInteger('encaissement_id')->nullable();
                 $table->decimal('montant_encaisse', 10, 2);
                 $table->string('mode_paiement');
                 $table->string('reference_recette')->unique();
                 $table->enum('type_recette', ['NORMAL', 'ADRA', 'TEAM3', 'DIVERS']);
                 $table->text('description')->nullable();
-                $table->unsignedBigInteger('created_by');
+                $table->unsignedInteger('created_by');
                 $table->string('year');
                 $table->text('observations')->nullable();
                 $table->timestamps();
                 $table->softDeletes();
 
-                // Index pour optimiser les requêtes
+                // Index
                 $table->index(['date_recette', 'year']);
                 $table->index(['student_id', 'year']);
                 $table->index(['class_id', 'year']);
@@ -43,7 +43,7 @@ class CreateRecettesTable extends Migration
                 $table->index(['mode_paiement', 'year']);
                 $table->index('encaissement_id');
 
-                // Contraintes de clés étrangères
+                // Foreign keys
                 $table->foreign('student_id')->references('id')->on('users')->onDelete('set null');
                 $table->foreign('class_id')->references('id')->on('my_classes')->onDelete('set null');
                 $table->foreign('payment_id')->references('id')->on('payments')->onDelete('set null');
@@ -52,7 +52,6 @@ class CreateRecettesTable extends Migration
                 $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
             });
         }
-        // If table exists, skip creation - indexes will be handled by Laravel automatically
     }
 
     /**
